+++ src/laser_pixelart.cpp
@@ -65,7 +65,7 @@ void Level::display_pixelart_lasers(point tl, point br)
     auto stop  = l.info.pixel_stop - l.info.pixel_map_at;
 
     auto   dist  = distance(start, stop);
-    float  steps = (int) ceil(dist) / TILE_WIDTH;
+    float  steps = (int) (ceil(dist) / (float) TILE_WIDTH);
     fpoint diff(stop.x - start.x, stop.y - start.y);
     fpoint step       = diff / steps;
     float  ninety_deg = RAD_360 / 4;
